* Halo1_BlamLib: build-cache-file rewrite progress * Halo1_BlamLib: Fixed cache_files_windows.cpp not compiling * Halo1_BlamLib: changed scenario's netgame_equipment block field to be declared non-const in runtime * Halo1_CheApe: s_main_globals can now be accessed

+++ OpenSauce/shared/Include/blamlib/Halo1/main/main.cpp
@@ -85,7 +85,7 @@ namespace Yelo
 
 		cstring main_get_map_name()					{ return GameState::MainGlobals()->scenario_tag_path; }
 
-		int32 PLATFORM_API main_get_campaign_level_from_name(cstring level_name)
+		int16 PLATFORM_API main_get_campaign_level_from_name(cstring level_name)
 		{
 			char name[128] = { };
